+++ docs/ru/keep-in-touch/release-notes.md
@@ -20,6 +20,7 @@
    * [sqrt(BigInt,Int,Int,Union)](/ru/ride/v6/functions/built-in-functions/math-functions#sqrtbigint)
 * Для встроенных функций строки [makeString](/ru/ride/v6/functions/built-in-functions/string-functions#makestring-list-string-string-string) и [split](/ru/ride/v6/functions/built-in-functions/string-functions#split-string-string-list-string) добавлены семейства аналогичных функций с различной сложностью в зависимости от размера данных. Если размер данных известен заранее, можно использовать более «дешевую» функцию.
 * Изменена сложность некоторых встроенных функций. Сложность всех функций представлена в разделе [[Ride v6] Встроенные функции](/ru/ride/v6/functions/built-in-functions/).
+* Изменено ограничение на второй параметр [функций строки](/ru/ride/v6/functions/built-in-functions/string-functions) и [функций массива байтов](/ru/ride/v6/functions/built-in-functions/byte-array-functions) `drop`, `dropRight`, `take`, `takeRight`.
 
 ### Node REST API